// src/lib/SortableTreeV2/types.ts
export type TreeItem<T> = {
	children?: TreeItem<T>[];
	id: UniqueIdentifier;
	/*
  Default: false.
   */
	collapsed?: boolean;

	/*
  If false, doesn't allow to drag&drop nodes so that they become children of current node.
  If you are showing files&directories it makes sense to set this to `true` for folders, and `false` for files.
  Default: true.
   */
	canHaveChildren?: boolean | ((dragItem: FlattenedItem<T>) => boolean);

	/*
  If true, the node can not be sorted/moved/dragged.
  Default: false.
   */
	disableSorting?: boolean;
} & T;
// ...
export type FlattenedItem<T> = {
	parentId: UniqueIdentifier | null;
	/*
  How deep in the tree is current item.
  0 - means the item is on the Root level,
  1 - item is child of Root level parent,
  etc.
   */
	depth: number;
	index: number;

	/*
  Is item the last one on it's deep level.
  This could be important for visualizing the depth level (e.g. in case of FolderTreeItemWrapper)
   */
	isLast: boolean;
	parent: FlattenedItem<T> | null;
} & TreeItem<T>;